+++ backend/canisters/local_user_index/impl/src/updates/mod.rs
@@ -2,6 +2,7 @@ pub mod bot_add_reaction;
 pub mod bot_create_channel;
 pub mod bot_delete_channel;
 pub mod bot_delete_messages;
+pub mod bot_remove_user;
 pub mod bot_send_message;
 pub mod bot_subscribe_to_events;
 pub mod c2c_create_community;